Dodan dio za novu kupovinu

+++ src/model/dto/PosjedujeMed.java
@@ -4,7 +4,7 @@ public class PosjedujeMed {
 	
 	private int IZVRCANI_MED_IdMeda;
 	private int PCELINJAK_IdPcelinjaka;
-	private int Kolicina;
+	private double Kolicina;
 	
 	public int getIZVRCANI_MED_IdMeda() {
 		return IZVRCANI_MED_IdMeda;
@@ -18,13 +18,13 @@ public class PosjedujeMed {
 	public void setPCELINJAK_IdPcelinjaka(int pCELINJAK_IdPcelinjaka) {
 		PCELINJAK_IdPcelinjaka = pCELINJAK_IdPcelinjaka;
 	}
-	public int getKolicina() {
+	public double getKolicina() {
 		return Kolicina;
 	}
-	public void setKolicina(int kolicina) {
+	public void setKolicina(double kolicina) {
 		Kolicina = kolicina;
 	}
-	public PosjedujeMed(int iZVRCANI_MED_IdMeda, int pCELINJAK_IdPcelinjaka, int kolicina) {
+	public PosjedujeMed(int iZVRCANI_MED_IdMeda, int pCELINJAK_IdPcelinjaka, double kolicina) {
 	
 		IZVRCANI_MED_IdMeda = iZVRCANI_MED_IdMeda;
 		PCELINJAK_IdPcelinjaka = pCELINJAK_IdPcelinjaka;
